requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
anthropic==0.44.0
python-dotenv==1.0.1
korean-lunar-calendar==0.3.1
